requirements.txt
colorlog==3.1.4
pymongo==3.7.1
Werkzeug==0.14.1
Flask==0.12.4
celery==4.2.1
flask_restplus==0.13.0
requests==2.19.1
dnspython==1.15.0
tld==0.9.1
PyYAML==5.1
geoip2==2.9.0
pyOpenSSL==19.0.0
mmh3==2.5.1
pyquery==1.4.1
openpyxl==3.0.0
gunicorn==19.9.0